# sRGB GPU formats, linear color-handling