--- a/libobs-d3d11/d3d11-subsystem.hpp
+++ b/libobs-d3d11/d3d11-subsystem.hpp
@@ -100,6 +100,78 @@ static inline DXGI_FORMAT ConvertGSTextureFormat(gs_color_format format)
 		return DXGI_FORMAT_BC3_UNORM;
 	case GS_R8G8:
 		return DXGI_FORMAT_R8G8_UNORM;
+	case GS_RGBA_SRGB:
+		return DXGI_FORMAT_R8G8B8A8_UNORM_SRGB;
+	case GS_BGRX_SRGB:
+		return DXGI_FORMAT_B8G8R8X8_UNORM_SRGB;
+	case GS_BGRA_SRGB:
+		return DXGI_FORMAT_B8G8R8A8_UNORM_SRGB;
+	}
+
+	return DXGI_FORMAT_UNKNOWN;
+}
+
+static inline DXGI_FORMAT ConvertGSTextureFormatResource(gs_color_format format)
+{
+	switch (format) {
+	case GS_UNKNOWN:
+		return DXGI_FORMAT_UNKNOWN;
+	case GS_A8:
+		return DXGI_FORMAT_A8_UNORM;
+	case GS_R8:
+		return DXGI_FORMAT_R8_UNORM;
+	case GS_RGBA:
+		return DXGI_FORMAT_R8G8B8A8_UNORM;
+	case GS_BGRX:
+		return DXGI_FORMAT_B8G8R8X8_UNORM;
+	case GS_BGRA:
+		return DXGI_FORMAT_B8G8R8A8_UNORM;
+	case GS_R10G10B10A2:
+		return DXGI_FORMAT_R10G10B10A2_UNORM;
+	case GS_RGBA16:
+		return DXGI_FORMAT_R16G16B16A16_UNORM;
+	case GS_R16:
+		return DXGI_FORMAT_R16_UNORM;
+	case GS_RGBA16F:
+		return DXGI_FORMAT_R16G16B16A16_FLOAT;
+	case GS_RGBA32F:
+		return DXGI_FORMAT_R32G32B32A32_FLOAT;
+	case GS_RG16F:
+		return DXGI_FORMAT_R16G16_FLOAT;
+	case GS_RG32F:
+		return DXGI_FORMAT_R32G32_FLOAT;
+	case GS_R16F:
+		return DXGI_FORMAT_R16_FLOAT;
+	case GS_R32F:
+		return DXGI_FORMAT_R32_FLOAT;
+	case GS_DXT1:
+		return DXGI_FORMAT_BC1_UNORM;
+	case GS_DXT3:
+		return DXGI_FORMAT_BC2_UNORM;
+	case GS_DXT5:
+		return DXGI_FORMAT_BC3_UNORM;
+	case GS_R8G8:
+		return DXGI_FORMAT_R8G8_UNORM;
+	case GS_RGBA_SRGB:
+		return DXGI_FORMAT_R8G8B8A8_TYPELESS;
+	case GS_BGRX_SRGB:
+		return DXGI_FORMAT_B8G8R8X8_TYPELESS;
+	case GS_BGRA_SRGB:
+		return DXGI_FORMAT_B8G8R8A8_TYPELESS;
+	}
+
+	return DXGI_FORMAT_UNKNOWN;
+}
+
+static inline DXGI_FORMAT ConvertGSTextureFormatNoSrgb(gs_color_format format)
+{
+	switch (format) {
+	case GS_RGBA_SRGB:
+		return DXGI_FORMAT_R8G8B8A8_UNORM;
+	case GS_BGRX_SRGB:
+		return DXGI_FORMAT_B8G8R8X8_UNORM;
+	case GS_BGRA_SRGB:
+		return DXGI_FORMAT_B8G8R8A8_UNORM;
 	}
 
 	return DXGI_FORMAT_UNKNOWN;
@@ -144,6 +216,12 @@ static inline gs_color_format ConvertDXGITextureFormat(DXGI_FORMAT format)
 		return GS_DXT3;
 	case DXGI_FORMAT_BC3_UNORM:
 		return GS_DXT5;
+	case DXGI_FORMAT_R8G8B8A8_UNORM_SRGB:
+		return GS_RGBA_SRGB;
+	case DXGI_FORMAT_B8G8R8X8_UNORM_SRGB:
+		return GS_BGRX_SRGB;
+	case DXGI_FORMAT_B8G8R8A8_UNORM_SRGB:
+		return GS_BGRA_SRGB;
 	}
 
 	return GS_UNKNOWN;
@@ -409,11 +487,14 @@ struct gs_texture : gs_obj {
 struct gs_texture_2d : gs_texture {
 	ComPtr<ID3D11Texture2D> texture;
 	ComPtr<ID3D11RenderTargetView> renderTarget[6];
+	ComPtr<ID3D11RenderTargetView> renderTargetNoSrgb[6];
 	ComPtr<IDXGISurface1> gdiSurface;
 
 	uint32_t width = 0, height = 0;
 	uint32_t flags = 0;
 	DXGI_FORMAT dxgiFormat = DXGI_FORMAT_UNKNOWN;
+	DXGI_FORMAT dxgiFormatResource = DXGI_FORMAT_UNKNOWN;
+	DXGI_FORMAT dxgiFormatNoSrgb = DXGI_FORMAT_UNKNOWN;
 	bool isRenderTarget = false;
 	bool isGDICompatible = false;
 	bool isDynamic = false;