--- a/libobs/obs-source-deinterlace.c
+++ b/libobs/obs-source-deinterlace.c
@@ -196,7 +196,7 @@ void set_deinterlace_texture_size(obs_source_t *source)
 {
 	if (source->async_gpu_conversion) {
 		source->async_prev_texrender =
-			gs_texrender_create(GS_BGRX, GS_ZS_NONE);
+			gs_texrender_create(GS_BGRX_SRGB, GS_ZS_NONE);
 
 		source->async_prev_texture = gs_texture_create(
 			source->async_convert_width,