--- a/libobs/obs.c
+++ b/libobs/obs.c
@@ -231,9 +231,9 @@ static bool obs_init_textures(struct obs_video_info *ovi)
 #endif
 	}
 
-	video->render_texture = gs_texture_create(ovi->base_width,
-						  ovi->base_height, GS_RGBA, 1,
-						  NULL, GS_RENDER_TARGET);
+	video->render_texture =
+		gs_texture_create(ovi->base_width, ovi->base_height,
+				  GS_RGBA_SRGB, 1, NULL, GS_RENDER_TARGET);
 
 	if (!video->render_texture)
 		return false;