--- a/plugins/obs-filters/chroma-key-filter.c
+++ b/plugins/obs-filters/chroma-key-filter.c
@@ -86,7 +86,7 @@ static inline void color_settings_update(struct chroma_key_filter_data *filter,
 	filter->brightness = (float)brightness;
 	filter->gamma = (float)gamma;
 
-	vec4_from_rgba(&filter->color, color);
+	vec4_from_rgba_srgb(&filter->color, color);
 }
 
 static inline void chroma_settings_update(struct chroma_key_filter_data *filter,
@@ -195,7 +195,7 @@ static void chroma_key_render(void *data, gs_effect_t *effect)
 	uint32_t height = obs_source_get_base_height(target);
 	struct vec2 pixel_size;
 
-	if (!obs_source_process_filter_begin(filter->context, GS_RGBA,
+	if (!obs_source_process_filter_begin(filter->context, GS_RGBA_SRGB,
 					     OBS_ALLOW_DIRECT_RENDERING))
 		return;