… plugins/obs-filters/color-correction-filter.c
… plugins/obs-filters/color-key-filter.c
… plugins/obs-filters/crop-filter.c
… plugins/obs-filters/data/luma_key_filter.effect
… plugins/obs-filters/gpu-delay.c
… plugins/obs-filters/luma-key-filter.c
… plugins/obs-filters/mask-filter.c
… plugins/obs-filters/scale-filter.c
… plugins/obs-filters/scroll-filter.c
… plugins/obs-filters/sharpness-filter.c
… plugins/win-capture/dc-capture.c