Starting up a draft PR for the implementation of sRGB formats to improve color handling. There are a few motivations:

- Some color blends are very wrong; e.g. the average of 0 and 255 is 188 for sRGB, not 128.
- Scale filters produce colors that are darker than they should be from interpolating nonlinear colors. Filtered texture samples have the same problem.
- Existing color filters probably produce images that are not mathematically correct although I'm still in the process of learning what correct means for some of them.
- Make color handling appear more consistent with #1713 to reduce "surprising" behavior. The same shaders can be used for both ACEScg and sRGB spaces in many situations since the shader values will be linear in both cases.

See this Google doc for more details as I continue to implement and test: https://docs.google.com/document/d/13Jh1QFgjveNKjdWXt3ksHzoAZl7ZS_Ich3QGeKnQzt4/edit?usp=sharing

**Implementation**
- [x] D3D support
- OpenGL support:
  - [x] WIndows
  - [x] Mac
  - [x] Linux
- Filters
  - [ ] chroma-key-filter
  - [ ] color-correction-filter
  - [ ] color-grade-filter
  - [ ] color-key-filter
  - [ ] crop-filter
  - [ ] gpu-delay
  - [ ] luma-key-filter [L* vs. Y'? (What do the two curves look like for 709 grays?)]
  - [ ] mask-filter
  - [ ] scale-filter
  - [ ] scroll-filter
  - [ ] sharpness-filter
- Inputs
  - [ ] linux-capture
  - [ ] obs-browser
  - [ ] obs-text
  - [ ] text-freetype2
  - [ ] win-capture
- Transitions
  - [ ] Cut
  - [ ] Fade
  - [ ] Fade to Color
  - [ ] Luma Wipe
  - [ ] Slide
  - [ ] Stinger
  - [ ] Swipe
- Image handling:
  - [x] FFmpeg
  - [ ] ImageMagick?
- [ ] decklink-output-ui
- [ ] [Other modified code paths]

**Testing**
- Document/test all code paths that have changed or are likely to regress here
  - Check OS/GPU/API permutations (Windows/Mac/Linux, NVIDIA/Intel, Direct3D/OpenGL)
- Check for Intel perf regressions
  - [x] Added sRGB shader math for Win32 HDC sources (measurements read cheaper, ha)
  - [x] General sRGB render target usage (definitely slower)
  - [ ] Upper bound on performance regression?
- [ ] TBD (All Tests)
  - [Add tests here]