@jpark37 2019-05-14
Did some quick GPA measurements on NVIDIA and Intel. NVIDIA/AMD should have fixed hardware to handle sRGB for free, but freaking Intel.

Here are some of the more noticeably affected draws in my test scene:
Crop filter: 178 us -> 306 us
Item into final mix: 185 us -> 423 us
Final mix into preview: 597 us -> 919 us

I found some Intel guidance to make sure it's the hardware.
https://software.intel.com/sites/default/files/managed/49/2b/Graphics-API-Performance-Guide-2.5.pdf
"Avoid using sRGB formats where unnecessary."

Is this a trade we're willing to make for baseline color correctness, or is this PR dead in the water?