@jp9000 2019-05-14
For OBS, where performance is so important, the most important rule to live by for a core changes like this is **any core feature that has a performance cost should almost always be made optional instead**.

Still, this code was worth testing just to see if there was a performance impact.  It's definitely unfortunate that there was though.  I was really hoping there wouldn't be.

As to whether we should make the tradeoff and merge it, the only thing I can say in response is that in almost seven years of working on this program, I have never really heard anyone complain or notice the issues that this PR or the other PR is fixing.  If I were to merge this, the sad fact is that almost no one would notice, and almost no one would care, except for the people whose performance had to be unwillingly sacrificed to make it possible.

My priority is always going to be performance over accuracy.  If the cost of increased accuracy is a hit to performance, then it must be an option, disabled by default, especially if we're attempting to solve something that almost no user and especially viewer really can reasonably notice, or would actually even care about.

I mean I am just not watching the countless streams/videos from OBS thinking "man if only we used sRGB, this wouldn't have happened".

But I do want an option for increased accuracy, because there are some people who value that.  Again however, it must be an option, and disabled by default.