@jpark37 2019-05-14
I hear where you're coming from. I was really hoping this would pan out to make #1713 less funky from a user's perspective. Let me mull this over for a while.