@jpark37 2019-05-14
Well, it looks like Intel will finally add support in Gen11:
https://software.intel.com/sites/default/files/managed/db/88/The-Architecture-of-Intel-Processor-Graphics-Gen11_R1new.pdf
"Gen11 enables two new optimizations to lossless color compression:"
"Support for sRGB surface formats for dynamic textures. Use of gamma corrected color
space is important especially as the usage of high dynamic range is increasing."

So Gen9 and prior has no support, Gen10 is cancelled, and Gen11 on will have native "free" sRGB support. The percentage of affected users should decrease over time.

I see 9.75% Intel usage on Steam HW survey. I'd be content to wait this out until the estimated user count is satisfactory.