@jpark37 2019-05-14
I suspect I'll still be around. Do you have criteria for how low you want the GPU share to go? Are hardware numbers available that might be more applicable than Steam? I can bump this PR, or recreate a new one once that threshold is reached.

From my own 1080p test scene w/preview of the same size, the overall frame time delta was 3.x ms to 4.x ms. (I can double-check when I get home.) The penalty scales with sRGB read/writes, i.e. more filters, but you had mentioned people don't usually use a lot of filters, so maybe this tops out around 2 ms under worse circumstances? That's probably still more than you'd want to give up on baseline even if it's just for integrated, so I'd be willing to wait. Putting this on a toggle, on top putting something like #1713/HDR on a toggle seems like a little much.