@mufunyo 2019-05-15
Honestly though, how many (serious) OBS users are using Intel HD to begin with? I've used it twice, and it has been awful both times. Once on a desktop i7 a few years ago where all the PCI-e slots were taken by capture cards, and I had to literally disable the preview to get 1080p30 without frame drops. The other time was trying to stream from my laptop, which, well, was a lost cause before I even started.

I have seen developers try to cater to Intel's awful architecture many times now (although this is the first instance where it's a multimedia application and not a game engine), and I really believe the only right way to deal with it is that where moving forward is concerned, is to not let yourself get held back by Intel's poor support. It all started with Intel 965 GMA, where Intel's engineers somehow managed to implement such a bare bones subset of PS 2.0 that Vista's Aero effects were just barely supported (and performing very poorly), and they have kept up these kind of half-implementations for years, frustrating developers in the process. By catering to them, you are enabling this behaviour to continue.

I do not believe that Intel HD users present a significant share of OBS' user base, and I am strongly of the opinion that persuading them to consider more powerful hardware will make them happier in the long run, because Intel HD GPUs just aren't up to the task of content creation. Using such an underpowered GPU should be at the user's own peril, and strongly discouraged.

Jim, I think at some point you need to make a decision whether OBS should just be this little toy you've created in 2012 to stream Starcraft II, or if it's a serious content creation tool that adheres to international video standards. Every piece of graphics software has had the "gamma problem" at some point in its lifetime, including Photoshop, Premiere and the Gimp, and it is a necessary growing pain that needs to be dealt with. Just because a large percentage of users are blind to incorrect colour processing doesn't mean it should be disabled by default. It should be *enabled* by default, and only disabled as a last resort to claw back performance on grossly underpowered systems.

Having said all this, I do think additional optimisation is possible. Back in my Avisynth days, I remember it was possible for filters to do compensate for gamma inside of the filter operation at very little extra cost. If we just leave things like scaling and texture sampling behind, a gamma corrected version of alpha blending, and a gamma corrected version of each of OBS' filters/effects, using standard UNORM surfaces should pose very little performance degradation. It would be a lot of work however to create and maintain two versions of each effect in the OBS code base. I'll leave in the middle whether or not that's worth the effort, especially since JPark has already put in so much diligent work.